feat : made hero component responsive
