--- components/hero/hero.css
@@ -1,4 +1,0 @@
-.hero-description {
-  font-size: 5rem;
-  font-weight: medium;
-}
