+++ index.html
@@ -238,7 +238,7 @@
 
     <div id="intro-container">
       <div class="hero">
-        <p>해외취업을 위한 커뮤니티 기반 알고리즘 스터디</p>      
+        <hero-component>해외취업을 위한 커뮤니티 기반 알고리즘 스터디</hero-component>
         <link-button size="big" variant="secondary" href="https://discord.gg/43UkheRV">디스코드 참여하기</link-button>
       </div>
 
@@ -252,14 +252,14 @@
         <img src="images/languages.png" alt="#" />
       </div>
       <div class="hero">
-        <p>지역에 관계없이 다양한 언어로 참여할 수 있어요</p>
+        <hero-component>지역에 관계없이 다양한 언어로 참여할 수 있어요</hero-component>
         <link-button size="big" variant="ghost" outline href="#join-instruction-container">참여방법 안내</link-button>
       </div>
     </div>
 
     <div id="review-container">
       <div class="hero">
-        <p>코드리뷰를 통해 새로운 관점을 배울 수 있어요</p>
+        <hero-component>코드리뷰를 통해 새로운 관점을 배울 수 있어요</hero-component>
         <link-button size="big" variant="secondary" href="https://discord.gg/43UkheRV">디스코드 참여하기</link-button>
       </div>
       <div class="review-image">
