+++ main.js
@@ -1,7 +1,7 @@
 import "./components/divider/divider.js";
 import "./components/footer-link/footer-link.js";
 import "./components/header.js"
-import "./components/hero/hero.js";
+import "./components/hero.js";
 import "./components/link-button/link-button.js";
 import "./components/review-item/review-item.js";
 import "./components/seo-meta-tag/seo-meta-tag.js";
